updated components and app structure
