+++ package.json
@@ -24,6 +24,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "qs": "^6.6.0"
+    "qs": "^6.6.0",
+    "react-router-dom": "^4.3.1"
   }
 }
