+++ src/Chart.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
+import './App.css';
+import Qs from 'qs';
+
+const Chart = (props => {
+    return (
+        <div>
+
+        </div>
+    )
+
+})
+
+export default Chart;
